# docs/swagger.yaml
basePath: /
definitions:
  core.ErrorDetail:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
  core.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/core.ErrorDetail'
    type: object
...
  entities.CreateEstadoCamionRequest:
    properties:
      camion_id:
        type: integer
      estado:
        type: string
      estado_id:
        type: integer
      observaciones:
        type: string
      timestamp:
        type: string
    required:
    - camion_id
    type: object
...
  entities.CreateNotificacionRequest:
    properties:
      activa:
        type: boolean
      creado_por:
        type: integer
      id_camion_relacionado:
        type: integer
      id_falla_relacionado:
        type: integer
      id_mantenimiento_relacionado:
        type: integer
      mensaje:
        type: string
      tipo:
        type: string
      titulo:
        type: string
      usuario_id:
        type: integer
    type: object
...
  entities.EstadoCamion:
    properties:
      camion_id:
        type: integer
      estado:
        description: nullable
        type: string
      estado_id:
        type: integer
      observaciones:
        description: nullable
        type: string
      timestamp:
        description: nullable
        type: string
    type: object
  entities.EstadoCamionListResponse:
    properties:
      code:
        type: integer
      data:
        items:
          $ref: '#/definitions/entities.EstadoCamion'
        type: array
      message:
        type: string
      success:
        type: boolean
    type: object
...
  entities.EstadoCamionResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/entities.EstadoCamion'
      message:
        type: string
      success:
        type: boolean
    type: object
...
  entities.Notificacion:
    properties:
      activa:
        type: boolean
      creado_por:
        type: integer
      created_at:
        type: string
      id_camion_relacionado:
        type: integer
      id_falla_relacionado:
        type: integer
      id_mantenimiento_relacionado:
        type: integer
      mensaje:
        type: string
      notificacion_id:
        type: integer
      tipo:
        type: string
      titulo:
        type: string
      usuario_id:
        type: integer
    type: object
...
  entities.NotificacionResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/entities.Notificacion'
      message:
        type: string
      success:
        type: boolean
    type: object
...
  entities.RegistroMantenimiento:
    properties:
      alerta_id:
        type: integer
      camion_id:
        type: integer
      coordinador_id:
        type: integer
      created_at:
        type: string
      fecha_realizada:
        type: string
      kilometraje_mantenimiento:
        type: number
      mecanico_responsable:
        type: string
      observaciones:
        type: string
      registro_id:
        type: integer
    type: object
...
  entities.RegistroMantenimientoMessageResponse:
    properties:
      code:
        type: integer
      message:
        type: string
      success:
        type: boolean
    type: object
  entities.RegistroMantenimientoResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/entities.RegistroMantenimiento'
      message:
        type: string
      success:
        type: boolean
    type: object
...
  entities.UpdateRegistroMantenimientoRequest:
    properties:
      alerta_id:
        type: integer
      camion_id:
        type: integer
      coordinador_id:
        type: integer
      fecha_realizada:
        type: string
      kilometraje_mantenimiento:
        type: number
      mecanico_responsable:
        type: string
      observaciones:
        type: string
    type: object
...
info:
  contact: {}
  description: API para gestión de recolección de residuos
  title: API Recolecta
  version: "1.0"
paths:
...
  /api/camion/:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.EstadoCamionListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Listar camiones
      tags:
      - Camion
    post:
      consumes:
      - application/json
      parameters:
      - description: Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entities.CreateEstadoCamionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entities.EstadoCamionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Crear camión
      tags:
      - Camion
...
  /api/notificaciones/falla/{falla_id}:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.NotificacionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Notificaciones por falla
      tags:
      - Notificacion
...
  /api/notificaciones/mantenimiento:
    post:
      parameters:
      - description: Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entities.CreateNotificacionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.NotificacionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Crear notificación de mantenimiento
      tags:
      - Notificacion
  /api/notificaciones/mantenimiento/{mantenimiento_id}:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.NotificacionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Notificaciones por mantenimiento
      tags:
      - Notificacion
...
  /api/registros-mantenimiento/{id}:
    delete:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.RegistroMantenimientoMessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Eliminar registro
      tags:
      - RegistroMantenimiento
    get:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.RegistroMantenimientoResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Registro por ID
      tags:
      - RegistroMantenimiento
    put:
      parameters:
      - description: Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entities.UpdateRegistroMantenimientoRequest'
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.RegistroMantenimientoResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Actualizar registro
      tags:
      - RegistroMantenimiento
...
  /api/reportes-mantenimiento-generado/{id}:
    delete:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.RegistroMantenimientoMessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Eliminar reporte generado
      tags:
      - ReporteMantenimientoGenerado
    get:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.RegistroMantenimientoResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Reporte generado por ID
      tags:
      - ReporteMantenimientoGenerado
    put:
      parameters:
      - description: Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entities.UpdateRegistroMantenimientoRequest'
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.RegistroMantenimientoResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Actualizar reporte generado
      tags:
      - ReporteMantenimientoGenerado
...
  /api/rutas/:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.EstadoCamionListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Listar rutas
      tags:
      - Ruta
    post:
      consumes:
      - application/json
      parameters:
      - description: Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entities.CreateEstadoCamionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entities.EstadoCamionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Crear ruta
      tags:
      - Ruta
...
securityDefinitions:
  BearerAuth:
    description: 'Escribe ''Bearer '' seguido de tu token JWT. Ejemplo: "Bearer eyJhbG..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"